feat: Convert to TypeScript and refactor CLI
The project has been fully converted to TypeScript. The `discogs-cover-cli.js` and `index.js` files are now deprecated stubs, and the CLI logic has been integrated into `index.ts`.

This change also includes:
- Removed explicit `process` import in favor of the global object for better type compatibility.
- Updated `package.json` to point the `bin` field to the compiled `index.js`.
- Updated `tsconfig.json` to include `.tsx` files and enable JSX.
- The CLI now uses `process.argv` for argument parsing and writes the output file.
