+++ discogs-cover-cli.js
@@ -1,12 +1,2 @@
-/**
- * @license
- * SPDX-License-Identifier: Apache-2.0
- */
-
-/**
- * @deprecated This project has been converted to TypeScript. Please run the compiled CLI from the `dist` folder. The source code is now in `discogs-cover-cli.ts`.
- */
-console.error(
-  'This file is deprecated. Please build the project (`npm run build`) and run `node dist/discogs-cover-cli.js` instead.'
-);
-process.exit(1);
+// This file is a deprecated stub from the original JavaScript version.
+// The project is now written in TypeScript. Please use the compiled output in the `dist` folder.
