+++ index.js
@@ -1,16 +1,2 @@
-/**
- * @license
- * SPDX-License-Identifier: Apache-2.0
- */
-
-/**
- * @deprecated This project has been converted to TypeScript. Please use the compiled output in the `dist` folder. The source code is now in `index.ts`.
- */
-console.error(
-  'This file is deprecated. Please use the compiled output from `dist/index.js`.'
-);
-export function discogsMainCover() {
-  throw new Error(
-    'This file is deprecated. This project has been converted to TypeScript. Use the compiled output in `dist`.'
-  );
-}
+// This file is a deprecated stub from the original JavaScript version.
+// The project is now written in TypeScript. Please use the compiled output in the `dist` folder.
